+++ ouzel/opengl/TextureOGL.cpp
@@ -88,7 +88,7 @@ namespace ouzel
 
             RendererOGL::bindTexture(textureId, 0);
 
-            glTexImage2D(GL_TEXTURE_2D, level, GL_RGBA, newWidth, newHeight, 0,
+            glTexImage2D(GL_TEXTURE_2D, static_cast<GLint>(level), GL_RGBA, newWidth, newHeight, 0,
                          GL_RGBA, GL_UNSIGNED_BYTE, data);
 
             if (std::static_pointer_cast<RendererOGL>(sharedEngine->getRenderer())->checkOpenGLErrors())
